rename pseudo_gen to pseudobulk
